First Chapter Finished

+++ database/migrations/2019_08_05_191354_create_answers_table.php
@@ -15,7 +15,7 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('question_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
             $table->text('content');
             $table->timestamps();
